feat: useCheckBox hook 구현
